hf-spaces-deployment/requirements.txt
# Universal LoRA Requirements for HF Spaces
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
pydantic==2.5.0
numpy==1.24.3

# PyTorch for Universal LoRA
torch==2.0.1
torchvision==0.15.2

# Diffusers for LoRA loading
diffusers==0.21.4
transformers==4.35.2

# LoRA support
peft==0.6.2
safetensors==0.4.0

# Hugging Face Hub
huggingface-hub==0.17.3

# Additional ML dependencies
accelerate==0.24.1